third_party/blink/web_tests/http/tests/local/blob/send-data-blob-expected.txt

Test for building blobs with multiple strings combined by BlobBuilder and sending them via XMLHttpRequest.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

* BlobBuilder.append(string)
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS 'type/foo' is 'type/foo'
PASS Expected response data received: OK
* BlobBuilder.append(blob)
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS 'type/bar' is 'type/bar'
PASS Expected response data received: OK
* BlobBuilder.append(string/blob)
PASS Expected response data received: OK
PASS Expected response data received: OK
* BlobBuilder.append(string/blob) - with recycled blob
PASS Expected response data received: OK
* BlobBuilder.append(string) - with line-endings
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS Expected response data received: OK
* BlobBuilder.append(string, "native") - with line-endings
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS Expected response data received: OK
PASS successfullyParsed is true

TEST COMPLETE
